+++ lib/whatsup_github/yaml_formatter.rb
